bl updater app
This builds a second binary that will be flashed like the main firmware
and will then update the bootloader, just like the st2. It additionaly
aligns the cmsis + config with the main app, and ensures the hw is
mostly completely initalized before the main app starts, which allows
skipping the init in the main fw.

Main changes:
* bl freq 12/48Mhz -> 120Mhz
* ext 32khz disabled
* "proper" init time for xosc
* no usb init unless DFU mode is actually used
* errata 2.13.1 clock init/lock fixups
* no guessing according to f++&%! bootprog flags where the main app is

Change-Id: I1403d35497cdea41ea5bed2c94b9e729ca352772