add fw update crcstub
Adds a crc stub that lives in front of the dfu bl updater app and checks
it.

The whole process is somewhat complicated to ensure we end up not just
with
a firmware file that contains the right crc for the right offsets, but
the elf
file is updated as well, to ensure the result is debuggable as-is.

The stub is built as part of the firmware code, and ends up in its own
section.
Linker script padding and alignment ensures that this works.

Change-Id: I39eae7aaafd5531db6ce48837c9499432caadbed